# `common/assets/process.go` - The first line of the function ‘extractPrefixTo’ uses a local variable ‘compressedFiles’. But I can’t find it and IDE reportes an error about it. # `go.mod` I have deleted useless additions. # `common/runtime/init.go` Moved `"github.com/turt2live/matrix-media-repo/storage/datastore/ds_gcp" to …datastore/ds_s3 We added a ‘GCP(Google Cloud Platform) support.md’.
