+++ GCP (Google Cloud Platform) support.md
@@ -1,0 +1,8 @@
+# GCP (Google Cloud Platform) support
+
+[There are four ways recommended by google to use gcp](https://cloud.google.com/docs/authentication)
+
+For now we add support for  `Service account key`. You need to create a service account, and download its private key as a JSON file then configure it in  your config file.
+
+[Tips about it](https://cloud.google.com/docs/authentication/production)
+
